blob is the encoded form of batch
