+++ aggregator/src/aggregation/barycentric.rs
@@ -349,7 +349,7 @@ pub fn interpolate(z: Scalar, coefficients: &[Scalar; BLOB_WIDTH]) -> Scalar {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::blob::{BlobData, KZG_TRUSTED_SETUP};
+    use crate::blob::{BatchData, KZG_TRUSTED_SETUP};
     use c_kzg::{Blob as RethBlob, KzgProof};
     use std::collections::BTreeSet;
 
@@ -386,7 +386,7 @@ mod tests {
 
     #[test]
     fn interpolate_matches_reth_implementation() {
-        let blob = BlobData::from(&vec![
+        let batch = BatchData::from(&vec![
             vec![30; 56],
             vec![200; 100],
             vec![0; 340],
@@ -396,8 +396,8 @@ mod tests {
         for z in 0..10 {
             let z = Scalar::from(u64::try_from(13241234 + z).unwrap());
             assert_eq!(
-                reth_point_evaluation(z, &blob.get_coefficients().map(|c| Scalar::from_raw(c.0))),
-                interpolate(z, &blob.get_coefficients().map(|c| Scalar::from_raw(c.0)))
+                reth_point_evaluation(z, &batch.get_coefficients().map(|c| Scalar::from_raw(c.0))),
+                interpolate(z, &batch.get_coefficients().map(|c| Scalar::from_raw(c.0)))
             );
         }
     }
