+++ aggregator/src/aggregation/decompression.rs
@@ -1,1 +1,3 @@
+pub(crate) mod decoder;
+
 mod witgen;
